fix(homepage): call MFL auction API directly, drop SSR self-fetch
The Current Auctions card never appeared on production even though the
visibility gate was true and there were active auctions. Root cause: the
homepage loader was self-fetching /api/live-auction during SSR
(`fetch(new URL('/api/live-auction', Astro.url))`). On Vercel the
serverless function calling its own public domain can fail silently
(timeout, no DNS, etc.); the catch returned [] and the card was hidden.
Confirmed in runtime logs — zero /api/live-auction hits despite the
homepage being rendered.

Fix:
- Extract the MFL fetch + parse logic into src/utils/live-auctions.ts
  (`fetchLiveAuctions`).
- /api/live-auction now thinly wraps that util — same JSON contract for
  the client-side 60s poller in HpAuctionsCard.
- The homepage loader (loadActiveAuctions) calls fetchLiveAuctions
  directly, removing the self-fetch entirely. Faster too — no extra
  hop and no JSON round-trip.
- Add a `[homepage-auctions] snapshot=N active=M` log so future failures
  surface in Vercel logs instead of silently hiding the card.
